alteracoes 08/11

+++ HealthyProject/Models/Utilizador.cs
@@ -27,7 +27,7 @@ namespace HealthyProject.Models
         public Nullable<System.DateTime> Data_nascimento { get; set; }
         public Nullable<int> Peso { get; set; }
         public Nullable<int> Altura { get; set; }
-        public Nullable<int> Actividade_fisica { get; set; }
+        public Nullable<double> Actividade_fisica { get; set; }
         public Nullable<int> Nr_horas_sono { get; set; }
         public Nullable<int> Nr_refeicoes { get; set; }
         public Nullable<bool> Habitos_alcoolicos { get; set; }
